Move Thank You Screens and ValidationStudyScreens (#502)
