+++ src/CovidApp.tsx
@@ -16,9 +16,9 @@ import { CountrySelectScreen } from '@covid/features/CountrySelectScreen';
 import { DrawerMenu } from '@covid/features/DrawerMenu';
 import { ScreenParamList } from '@covid/features/ScreenParamList';
 import { SplashScreen } from '@covid/features/SplashScreen';
-import ThankYouScreen from '@covid/features/ThankYouScreen';
-import ThankYouUKScreen from '@covid/features/ThankYouUKScreen';
-import ViralThankYouScreen from '@covid/features/ViralThankYouScreen';
+import ThankYouScreen from '@covid/features/thank-you/ThankYouScreen';
+import ThankYouUKScreen from '@covid/features/thank-you/ThankYouUKScreen';
+import ViralThankYouScreen from '@covid/features/thank-you/ViralThankYouScreen';
 import CovidTestDetailScreen from '@covid/features/covid-tests/CovidTestDetailScreen';
 import DescribeSymptomsScreen from '@covid/features/assessment/DescribeSymptomsScreen';
 import HealthWorkerExposureScreen from '@covid/features/assessment/HealthWorkerExposureScreen';
@@ -49,9 +49,9 @@ import Welcome1Screen from '@covid/features/register/Welcome1Screen';
 import Welcome2Screen from '@covid/features/register/Welcome2Screen';
 import { WelcomeRepeatScreen } from '@covid/features/register/WelcomeRepeatScreen';
 import { PrivacyPolicyUKScreen } from '@covid/features/register/gb/PrivacyPolicyUKScreen';
-import ValidationStudyConsentScreen from '@covid/features/register/gb/ValidationStudyConsentScreen';
-import ValidationStudyInfoScreen from '@covid/features/register/gb/ValidationStudyInfoScreen';
-import ValidationStudyIntroScreen from '@covid/features/register/gb/ValidationStudyIntroScreen';
+import ValidationStudyConsentScreen from '@covid/features/validation-study/ValidationStudyConsentScreen';
+import ValidationStudyInfoScreen from '@covid/features/validation-study/ValidationStudyInfoScreen';
+import ValidationStudyIntroScreen from '@covid/features/validation-study/ValidationStudyIntroScreen';
 import PrivacyPolicySVScreen from '@covid/features/register/sv/PrivacyPolicySVScreen';
 import BeforeWeStartUS from '@covid/features/register/us/BeforeWeStartUS';
 import { NursesConsentUSScreen } from '@covid/features/register/us/NursesConsentUS';
