+++ src/features/validation-study/ValidationStudyIntroScreen.tsx
@@ -13,8 +13,8 @@ import { Header } from '@covid/components/Screen';
 import { lazyInject } from '@covid/provider/services';
 import { Services } from '@covid/provider/services.types';
 
-import appCoordinator from '../../AppCoordinator';
-import { ScreenParamList } from '../../ScreenParamList';
+import appCoordinator from '../AppCoordinator';
+import { ScreenParamList } from '../ScreenParamList';
 
 type Props = {
   navigation: StackNavigationProp<ScreenParamList, 'ValidationStudyIntro'>;
